comm_split_type: split neighborhood logic into separate file
